# Improve `cupy.argwhere`