@leofang 2020-11-19
Thanks, @anaruse! I was also looking at this recently as it's closely related to #3553. Should we add the support of CUB's device scan in this PR, or do it separately?