@anaruse 2020-11-20
Thanks for your comment, @leofang.

It would be great if we could further improve the performance of `argwhere` by use of CUB. However, this PR uses "imcomplete scan", a method that does not complete the scan process, so I don't think we can use the CUB's device scan in this PR.

If you use the CUB's block scan, I think there is a possibility to further improve the performance in combination with this PR.