@leofang 2020-11-20
Hi @anaruse Sorry I wasn't organizing my thoughts well. When I saw you mentioned
>  the execution time of an inclusive scan has to be reduced

what I had in mind was simply that the inclusive scan (a complete one) could perhaps be replaced by CUB's device scan. I could try CUB's block scan too, if there's an interest, but I don't want that to be a blocker 😛