@emcastillo 2020-11-20
Can you please detail a bit the changes that you did to the non-zero kernel? I can't understand the logic behind it :).

Thanks!