@anaruse 2020-11-20
> Can you please detail a bit the changes that you did to the non-zero kernel? I can't understand the logic behind it :).

The basic idea is to skip the execution of `kern_add`, which is performed at the end of the inclusive scan, and do the equivalent computation in `nonzero_kernel` of `argwhere` instead, that would have been done in the `kern_add`.

The advantage of this method is that
* No need to compute all the elements, and
* No need to write the results back to device memory.