@anaruse 2020-12-11
Closing this, as https://github.com/cupy/cupy/pull/4367 is better.